Fixing bulk training issues
made the model trainer default to tail as it was attempting to get 1
year of data for each symbol current year. There were files that had
multiple years of data but seemed to halt trading before recent
dates. This was breaking.

Also trimmed the cache listing to 5 years of data only. This is because
I want to train on 2 years of data and if you only have 2 years of data in
aggregate our trend caculations start to contain nans which blows up change
detection.